前期策划工作计划

+++ Ft-Cloud/ft-services/ft-service-pm/src/main/java/com/hhwy/pm/qqch/qqchWorkPlan/service/IQqchWorkPlanService.java
@@ -24,7 +24,7 @@ public interface IQqchWorkPlanService {
 
     List<QqchWorkPlan> getQqchWorkPlanList(QqchWorkPlan qqchWorkPlan);
 
-    int insertQqchWorkPlan(QqchWorkPlan qqchWorkPlan);
+    Long insertQqchWorkPlan(QqchWorkPlan qqchWorkPlan);
 
     int insertQqchWorkPlanList(List<QqchWorkPlan> qqchWorkPlanList);
